+++ dico_toolbox/__init__.py
@@ -1,2 +1,5 @@
 from . import aims_tools 
 from . import transform
+from . import data
+from . import wrappers
+from . import convert
